refactor: make main and tools private
