+++ fglatch/_tools/__init__.py
@@ -1,0 +1,5 @@
+from fglatch._tools.submit import submit
+
+__all__ = [
+    "submit",
+]
