--- fglatch/tools/__init__.py
@@ -1,5 +1,0 @@
-from fglatch.tools.submit import submit
-
-__all__ = [
-    "submit",
-]
